max likelihood naive bayes
this will searve as a baseline to check against the fully bayesian
implementation. a few things to note. I am only getting 50% accuracy
(under 0-1 metric). this could be a mix of many things:

1. the feature engineering is super basic. better performance might be
obtained by providing the sklearn vectorizer with an explicit vocab
list.

2. there is underflow happening when computing the predictive density.
   going to try to switch to log probabilities. should be a good
learning experience

the goal is to get accuracy over 70%